Added functionality to consume items command: cons
Items can select between any attribute (that we allow) and then modify
that as they see fit. Currently the Milk removes 1 hp and the
HealthPotion adds 10.

Also created the vitals command which currently just shows the health of
the player.

I need to really start consolidating the commands down. Maybe picking a
lenght that all of the commands will be, as well as how to handle two
word things like 'health potion' vs 'HealthPotion'

Thats a problem for future me